Fixed JS Light Theme bug
I had the function set up to not account for when a user clicks on dark mode, then tries to go back to light mode. A bug would happen where it wouldn't display the bright mode in certain sections after re-clicking light mode. Just simply added the same code that changed it to dark mode, but back to light mode after the if statement was read.